+++ aim/engine/configs.py
@@ -5,6 +5,7 @@ AIM_ANNOT_DIR_NAME = 'annotations'
 AIM_MEDIA_DIR_NAME = 'media'
 AIM_IMAGES_DIR_NAME = 'images'
 AIM_MODELS_DIR_NAME = 'models'
+AIM_CORR_DIR_NAME = 'correlations'
 AIM_CONFIG_FILE_NAME = 'config.json'
 AIM_DEFAULT_BRANCH_NAME = 'default'