+++ aim/sdk/artifacts/__init__.py
@@ -2,3 +2,4 @@ from aim.sdk.artifacts.metric import *
 from aim.sdk.artifacts.media import *
 from aim.sdk.artifacts.misclassification import *
 from aim.sdk.artifacts.model import *
+from aim.sdk.artifacts.correlation import *