+++ aim/version_control/git.py
@@ -1,6 +1,6 @@
 from aim.version_control.base import Base
 
 
-class Git(Base):
+class GitAdapter(Base):
     def get_diff(self):
-        ...
+        print('Git diff')