+++ aim/version_control/base.py
@@ -2,12 +2,17 @@ from abc import ABC, abstractmethod
 
 
 class Base(ABC):
+    @staticmethod
     @abstractmethod
-    def commit(self):
+    def get_repo():
         ...
 
     @abstractmethod
-    def checkout(self):
+    def get_untracked_files(self):
+        ...
+
+    @abstractmethod
+    def commit_changes_to_branch(self, commit_msg, branch_name, branch_prefix):
         ...
 
     @abstractmethod