+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add aim correlation tracking (gorarakelyan)
+
 ## 0.2.6 Nov 5 2019
 - Update aim track logger (gorarakelyan)